Trekk ut sortering og tidslinjerad-logikk til egne filer
