+++ src/components/TidslinjeKombinert.tsx
@@ -1,67 +1,29 @@
 import React, { useState } from 'react'
-import { StethoscopeIcon, TasklistIcon, SplitHorizontalIcon, EarthIcon } from '@navikt/aksel-icons'
 import { BodyShort, Box, Timeline } from '@navikt/ds-react'
 
 import { KlippetSykepengesoknadRecord, Soknad, dayjsToDate } from '../queryhooks/useSoknader'
 import type { Sykmelding } from '../queryhooks/useSykmeldinger'
-import gruppertOgFiltrert, { ArbeidsgiverGruppering, SoknadGruppering } from '../utils/gruppering'
+import gruppertOgFiltrert from '../utils/gruppering'
 import { filtrerPaFilter } from '../utils/filterlogikk'
 import { hentDatospenn, validerSykmeldingsDatoer } from '../utils/sykmeldingValidering'
 import { beregnAktivTidsvindu, erPeriodeInnenforTidsvindu } from '../utils/tidslinjeUtils'
-import { arbeidsgiverLabelForSoknader } from '../utils/soknadArbeidsgiverLabel'
 
 import { Filter, ValgteFilter } from './Filter'
 import VelgZoomPeriode from './VelgZoomPeriode'
-import { timelinePeriodeStatus } from './soknad/Tidslinje'
 import {
     grupperSykmeldingerPaArbeidsgiver,
     perioderMedDatoer,
     sorterPerioder,
-    sykmeldingStatus,
-    SykmeldingerPerArbeidsgiver,
 } from './sykmelding/sykmeldingTidslinjeUtils'
-import ViktigeFeltForSoknad from './periodeinfo/ViktigeFeltForSoknad'
-import ViktigeFeltForSykmelding from './periodeinfo/ViktigeFeltForSykmelding'
-import DetaljerDrawer, {
-    DrawerInnhold,
-    lagKlippetSoknadDrawerInnhold,
-    lagSykmeldingDrawerInnhold,
-    lagSoknadDrawerInnhold,
-    lagOppholdUtlandSoknadDrawerInnhold,
-} from './DetaljerDrawer'
+import DetaljerDrawer, { DrawerInnhold } from './DetaljerDrawer'
+import { lagSykmeldingRader } from './kombinert/SykmeldingRader'
+import { lagSoknadRader } from './kombinert/SoknadRader'
+import { lagOppholdUtlandPins } from './kombinert/OppholdUtlandPins'
 
 interface Props {
     sykmeldinger: Sykmelding[]
     soknader: Soknad[]
     klipp: KlippetSykepengesoknadRecord[]
-}
-
-function sorterSykmeldingGrupperEtterSignaturDato(
-    entries: [string, SykmeldingerPerArbeidsgiver][],
-): [string, SykmeldingerPerArbeidsgiver][] {
-    return entries.slice().sort((a, b) => {
-        const maxA = Math.max(...a[1].sykmeldinger.map((s) => s.signaturDato?.valueOf() ?? 0))
-        const maxB = Math.max(...b[1].sykmeldinger.map((s) => s.signaturDato?.valueOf() ?? 0))
-        return maxB - maxA
-    })
-}
-
-function sorterSoknadGrupperEtterSignaturDato(
-    entries: [string, ArbeidsgiverGruppering][],
-): [string, ArbeidsgiverGruppering][] {
-    return entries.slice().sort((a, b) => {
-        const maxA = Math.max(
-            ...[...a[1].sykmeldinger.values()].flatMap((syk) =>
-                [...syk.soknader.values()].map((sg) => sg.soknad.sykmeldingUtskrevet?.valueOf() ?? 0),
-            ),
-        )
-        const maxB = Math.max(
-            ...[...b[1].sykmeldinger.values()].flatMap((syk) =>
-                [...syk.soknader.values()].map((sg) => sg.soknad.sykmeldingUtskrevet?.valueOf() ?? 0),
-            ),
-        )
-        return maxB - maxA
-    })
 }
 
 export default function TidslinjeKombinert({ sykmeldinger, soknader, klipp }: Props) {
@@ -126,253 +88,16 @@ export default function TidslinjeKombinert({ sykmeldinger, soknader, klipp }: Pr
           ).length
         : filtrerteSoknaderAntall
 
-    const sykmeldingRader = aktivTidsvindu
-        ? sorterSykmeldingGrupperEtterSignaturDato(Array.from(sykmeldingerGruppertPaArbeidsgiver.entries())).flatMap(
-              ([arbeidsgiverId, arbeidsgiver]) => {
-                  const perioder_med_innhold = arbeidsgiver.sykmeldinger.flatMap((sykmelding) => {
-                      const status = sykmeldingStatus(sykmelding.sykmeldingStatus?.statusEvent)
-                      const perioder = sorterPerioder(perioderMedDatoer(sykmelding))
+    const handlePeriodeValgt = (periodeId: string | null, kildeId: string | null, drawer: DrawerInnhold | null) => {
+        setAktivPeriodeId(periodeId)
+        setAktivDrawerKildeId(kildeId)
+        setDrawerInnhold(drawer)
+    }
 
-                      if (perioder.length === 0) return []
-
-                      const forstePeriode = perioder[0]
-                      const sistePeriode = perioder[perioder.length - 1]
-
-                      if (
-                          !erPeriodeInnenforTidsvindu(
-                              forstePeriode.startDato,
-                              sistePeriode.sluttDato,
-                              aktivTidsvindu.fra,
-                              aktivTidsvindu.til,
-                          )
-                      ) {
-                          return []
-                      }
-
-                      const harFlerePerioder = perioder.length > 1
-                      const ikon = harFlerePerioder ? <SplitHorizontalIcon aria-hidden /> : undefined
-                      const periodeKey = `${sykmelding.id}-${forstePeriode.fom}-${sistePeriode.tom}`
-
-                      const periodeInfo = <ViktigeFeltForSykmelding sykmelding={sykmelding} perioder={perioder} />
-
-                      const sykmeldingAktivId = sykmelding.id
-
-                      return [
-                          <Timeline.Period
-                              start={forstePeriode.startDato}
-                              end={sistePeriode.sluttDato}
-                              status={status}
-                              icon={ikon}
-                              className="ring-1 ring-inset ring-white/95"
-                              key={periodeKey}
-                              isActive={aktivPeriodeId === sykmeldingAktivId}
-                              onSelectPeriod={() => {
-                                  if (aktivDrawerKildeId === sykmeldingAktivId) {
-                                      setAktivPeriodeId(null)
-                                      setAktivDrawerKildeId(null)
-                                      setDrawerInnhold(null)
-                                  } else {
-                                      setAktivPeriodeId(sykmeldingAktivId)
-                                      setAktivDrawerKildeId(sykmeldingAktivId)
-                                      setDrawerInnhold(lagSykmeldingDrawerInnhold(sykmelding, periodeInfo))
-                                  }
-                              }}
-                          />,
-                      ]
-                  })
-
-                  if (perioder_med_innhold.length === 0) return []
-
-                  return [
-                      <Timeline.Row
-                          key={`syk-${arbeidsgiverId}`}
-                          label={arbeidsgiver.label}
-                          icon={<StethoscopeIcon aria-hidden fontSize="1.5rem" />}
-                      >
-                          {perioder_med_innhold}
-                      </Timeline.Row>,
-                  ]
-              },
-          )
-        : []
-
-    const soknadRader = aktivTidsvindu
-        ? sorterSoknadGrupperEtterSignaturDato(Array.from(soknaderGruppert.entries())).flatMap(([arbId, arb]) => {
-              if (arbId === 'opphold_utland') return []
-
-              const label = arbeidsgiverLabelForSoknader(arbId, arb, soknaderGruppert)
-
-              const perioder_med_innhold = Array.from(arb.sykmeldinger.entries()).flatMap(([sykId, syk]) => {
-                  const erGhostSykmelding = sykId.endsWith('_GHOST')
-
-                  return Array.from(syk.soknader.values())
-                      .flatMap((sok: SoknadGruppering) => {
-                          const klippingAvSoknad = sok.klippingAvSoknad
-                              .filter((k) => {
-                                  const fom = dayjsToDate(k.fom)
-                                  const tom = dayjsToDate(k.tom)
-                                  return (
-                                      fom &&
-                                      tom &&
-                                      erPeriodeInnenforTidsvindu(fom, tom, aktivTidsvindu.fra, aktivTidsvindu.til)
-                                  )
-                              })
-                              .map((k) => {
-                                  const sykmeldingId = k.sykmeldingUuid ?? null
-                                  const erAktiv = aktivPeriodeId !== null && aktivPeriodeId === sykmeldingId
-                                  const kildeId = k.id
-
-                                  return (
-                                      <Timeline.Period
-                                          start={dayjsToDate(k.fom)!}
-                                          end={dayjsToDate(k.tom)!}
-                                          status="neutral"
-                                          key={k.tom}
-                                          isActive={erAktiv}
-                                          onSelectPeriod={() => {
-                                              if (aktivDrawerKildeId === kildeId) {
-                                                  setAktivPeriodeId(null)
-                                                  setAktivDrawerKildeId(null)
-                                                  setDrawerInnhold(null)
-                                              } else {
-                                                  setAktivPeriodeId(sykmeldingId)
-                                                  setAktivDrawerKildeId(kildeId)
-                                                  setDrawerInnhold(lagKlippetSoknadDrawerInnhold(k))
-                                              }
-                                          }}
-                                      />
-                                  )
-                              })
-
-                          if (!erGhostSykmelding) {
-                              const sokFom = dayjsToDate(sok.soknad.fom!)
-                              const sokTom = dayjsToDate(sok.soknad.tom!)
-                              if (
-                                  sokFom &&
-                                  sokTom &&
-                                  erPeriodeInnenforTidsvindu(sokFom, sokTom, aktivTidsvindu.fra, aktivTidsvindu.til)
-                              ) {
-                                  const sykmeldingId = sok.soknad.sykmeldingId ?? null
-                                  const erAktiv = aktivPeriodeId !== null && aktivPeriodeId === sykmeldingId
-                                  const kildeId = sok.soknad.id
-                                  klippingAvSoknad.push(
-                                      <Timeline.Period
-                                          start={sokFom}
-                                          end={sokTom}
-                                          status={timelinePeriodeStatus(sok.soknad.status)}
-                                          key={sok.soknad.tom?.toISOString() ?? sok.soknad.id}
-                                          isActive={erAktiv}
-                                          onSelectPeriod={() => {
-                                              if (aktivDrawerKildeId === kildeId) {
-                                                  setAktivPeriodeId(null)
-                                                  setAktivDrawerKildeId(null)
-                                                  setDrawerInnhold(null)
-                                              } else {
-                                                  setAktivPeriodeId(sykmeldingId)
-                                                  setAktivDrawerKildeId(kildeId)
-                                                  setDrawerInnhold(
-                                                      lagSoknadDrawerInnhold(
-                                                          sok.soknad,
-                                                          <ViktigeFeltForSoknad soknad={sok.soknad} />,
-                                                      ),
-                                                  )
-                                              }
-                                          }}
-                                      />,
-                                  )
-                              }
-                          }
-
-                          return klippingAvSoknad
-                      })
-                      .concat(
-                          syk.klippingAvSykmelding
-                              .filter((k) => {
-                                  const fom = dayjsToDate(k.fom)
-                                  const tom = dayjsToDate(k.tom)
-                                  return (
-                                      fom &&
-                                      tom &&
-                                      erPeriodeInnenforTidsvindu(fom, tom, aktivTidsvindu.fra, aktivTidsvindu.til)
-                                  )
-                              })
-                              .map((k) => {
-                                  const sykmeldingId = k.sykmeldingUuid ?? null
-                                  const erAktiv = aktivPeriodeId !== null && aktivPeriodeId === sykmeldingId
-                                  const kildeId = k.id
-
-                                  return (
-                                      <Timeline.Period
-                                          start={dayjsToDate(k.fom)!}
-                                          end={dayjsToDate(k.tom)!}
-                                          status="neutral"
-                                          key={k.tom}
-                                          isActive={erAktiv}
-                                          onSelectPeriod={() => {
-                                              if (aktivDrawerKildeId === kildeId) {
-                                                  setAktivPeriodeId(null)
-                                                  setAktivDrawerKildeId(null)
-                                                  setDrawerInnhold(null)
-                                              } else {
-                                                  setAktivPeriodeId(sykmeldingId)
-                                                  setAktivDrawerKildeId(kildeId)
-                                                  setDrawerInnhold(lagKlippetSoknadDrawerInnhold(k))
-                                              }
-                                          }}
-                                      />
-                                  )
-                              }),
-                      )
-              })
-
-              if (perioder_med_innhold.length === 0) return []
-
-              return [
-                  <Timeline.Row
-                      key={`sok-${arbId}`}
-                      label={label}
-                      icon={<TasklistIcon aria-hidden fontSize="1.5rem" />}
-                  >
-                      {perioder_med_innhold}
-                  </Timeline.Row>,
-              ]
-          })
-        : []
-
-    const oppholdUtlandPins = aktivTidsvindu
-        ? Array.from(soknaderGruppert.get('opphold_utland')?.sykmeldinger.values() ?? []).flatMap((syk) =>
-              Array.from(syk.soknader.values()).flatMap((sok) => {
-                  const dato = dayjsToDate(sok.soknad.opprettetDato)
-                  if (!dato) return []
-                  const kildeId = sok.soknad.id
-                  return [
-                      <Timeline.Pin
-                          key={sok.soknad.id}
-                          date={dato}
-                          onClick={() => {
-                              if (aktivDrawerKildeId === kildeId) {
-                                  setAktivDrawerKildeId(null)
-                                  setDrawerInnhold(null)
-                              } else {
-                                  setAktivDrawerKildeId(kildeId)
-                                  setDrawerInnhold(
-                                      lagOppholdUtlandSoknadDrawerInnhold(
-                                          sok.soknad,
-                                          <ViktigeFeltForSoknad soknad={sok.soknad} />,
-                                      ),
-                                  )
-                              }
-                          }}
-                      >
-                          <span className="flex items-center gap-1 text-sm">
-                              <EarthIcon aria-hidden fontSize="1.25rem" />
-                              Opphold utland søknad
-                          </span>
-                      </Timeline.Pin>,
-                  ]
-              }),
-          )
-        : []
+    const handleDrawerValgt = (kildeId: string | null, drawer: DrawerInnhold | null) => {
+        setAktivDrawerKildeId(kildeId)
+        setDrawerInnhold(drawer)
+    }
 
     const labelKlasse = 'kombinert-tidslinje-boks'
 
@@ -396,7 +121,13 @@ export default function TidslinjeKombinert({ sykmeldinger, soknader, klipp }: Pr
                             startDate={aktivTidsvindu.fra}
                             key={`syk-${aktivTidsvindu.fra.toISOString()}-${aktivTidsvindu.til.toISOString()}`}
                         >
-                            {sykmeldingRader}
+                            {lagSykmeldingRader({
+                                sykmeldingerGruppertPaArbeidsgiver,
+                                aktivTidsvindu,
+                                aktivPeriodeId,
+                                aktivDrawerKildeId,
+                                onPeriodeValgt: handlePeriodeValgt,
+                            })}
                         </Timeline>
                     </Box>
                     <Box
@@ -412,8 +143,18 @@ export default function TidslinjeKombinert({ sykmeldinger, soknader, klipp }: Pr
                             startDate={aktivTidsvindu.fra}
                             key={`sok-${aktivTidsvindu.fra.toISOString()}-${aktivTidsvindu.til.toISOString()}`}
                         >
-                            {soknadRader}
-                            {oppholdUtlandPins}
+                            {lagSoknadRader({
+                                soknaderGruppert,
+                                aktivTidsvindu,
+                                aktivPeriodeId,
+                                aktivDrawerKildeId,
+                                onPeriodeValgt: handlePeriodeValgt,
+                            })}
+                            {lagOppholdUtlandPins({
+                                soknaderGruppert,
+                                aktivDrawerKildeId,
+                                onDrawerValgt: handleDrawerValgt,
+                            })}
                         </Timeline>
                     </Box>
                 </>
