+++ src/components/kombinert/OppholdUtlandPins.tsx
@@ -1,0 +1,51 @@
+import React from 'react'
+import { EarthIcon } from '@navikt/aksel-icons'
+import { Timeline } from '@navikt/ds-react'
+
+import { ArbeidsgiverGruppering } from '../../utils/gruppering'
+import { dayjsToDate } from '../../queryhooks/useSoknader'
+import { lagOppholdUtlandSoknadDrawerInnhold, type DrawerInnhold } from '../DetaljerDrawer'
+import ViktigeFeltForSoknad from '../periodeinfo/ViktigeFeltForSoknad'
+
+export type OnDrawerValgt = (kildeId: string | null, drawer: DrawerInnhold | null) => void
+
+interface Props {
+    soknaderGruppert: Map<string, ArbeidsgiverGruppering>
+    aktivDrawerKildeId: string | null
+    onDrawerValgt: OnDrawerValgt
+}
+
+export function lagOppholdUtlandPins({ soknaderGruppert, aktivDrawerKildeId, onDrawerValgt }: Props) {
+    return Array.from(soknaderGruppert.get('opphold_utland')?.sykmeldinger.values() ?? []).flatMap((syk) =>
+        Array.from(syk.soknader.values()).flatMap((sok) => {
+            const dato = dayjsToDate(sok.soknad.opprettetDato)
+            if (!dato) return []
+            const kildeId = sok.soknad.id
+
+            return [
+                <Timeline.Pin
+                    key={sok.soknad.id}
+                    date={dato}
+                    onClick={() => {
+                        if (aktivDrawerKildeId === kildeId) {
+                            onDrawerValgt(null, null)
+                        } else {
+                            onDrawerValgt(
+                                kildeId,
+                                lagOppholdUtlandSoknadDrawerInnhold(
+                                    sok.soknad,
+                                    <ViktigeFeltForSoknad soknad={sok.soknad} />,
+                                ),
+                            )
+                        }
+                    }}
+                >
+                    <span className="flex items-center gap-1 text-sm">
+                        <EarthIcon aria-hidden fontSize="1.25rem" />
+                        Opphold utland søknad
+                    </span>
+                </Timeline.Pin>,
+            ]
+        }),
+    )
+}
