+++ src/utils/kombinerTidslinjeSortering.test.ts
@@ -1,0 +1,93 @@
+import { describe, expect, it } from 'vitest'
+import dayjs from 'dayjs'
+
+import { Soknad } from '../queryhooks/useSoknader'
+import type { Sykmelding } from '../queryhooks/useSykmeldinger'
+import type { SykmeldingerPerArbeidsgiver } from '../components/sykmelding/sykmeldingTidslinjeUtils'
+
+import type { ArbeidsgiverGruppering } from './gruppering'
+import {
+    sorterSykmeldingGrupperEtterSignaturDato,
+    sorterSoknadGrupperEtterSignaturDato,
+} from './kombinerTidslinjeSortering'
+
+const lagSykGruppe = (id: string, signaturDato: string | null): [string, SykmeldingerPerArbeidsgiver] => [
+    id,
+    {
+        label: id,
+        dagNokler: new Set(),
+        sykmeldinger: [{ signaturDato: signaturDato ? dayjs(signaturDato) : undefined } as unknown as Sykmelding],
+    },
+]
+
+const lagSokGruppe = (id: string, sykmeldingUtskrevet: string | null): [string, ArbeidsgiverGruppering] => {
+    const soknad = new Soknad({
+        id: `soknad-${id}`,
+        soknadstype: 'SELVSTENDIGE_OG_FRILANSERE',
+        status: 'SENDT',
+        soknadPerioder: [],
+        sykmeldingUtskrevet: sykmeldingUtskrevet ?? undefined,
+    })
+    return [
+        id,
+        {
+            sykmeldinger: new Map([
+                [
+                    `sykmelding-${id}`,
+                    {
+                        soknader: new Map([[soknad.id, { soknad, klippingAvSoknad: [] }]]),
+                        klippingAvSykmelding: [],
+                    },
+                ],
+            ]),
+        },
+    ]
+}
+
+describe('sorterSykmeldingGrupperEtterSignaturDato', () => {
+    it('sorterer nyeste signaturDato øverst', () => {
+        const entries = [lagSykGruppe('gammel', '2026-01-01'), lagSykGruppe('ny', '2026-03-01')]
+
+        const resultat = sorterSykmeldingGrupperEtterSignaturDato(entries)
+
+        expect(resultat[0][0]).toBe('ny')
+        expect(resultat[1][0]).toBe('gammel')
+    })
+
+    it('plasserer gruppe uten signaturDato sist', () => {
+        const entries = [lagSykGruppe('uten-dato', null), lagSykGruppe('med-dato', '2026-02-01')]
+
+        const resultat = sorterSykmeldingGrupperEtterSignaturDato(entries)
+
+        expect(resultat[0][0]).toBe('med-dato')
+        expect(resultat[1][0]).toBe('uten-dato')
+    })
+
+    it('returnerer tom liste uendret', () => {
+        expect(sorterSykmeldingGrupperEtterSignaturDato([])).toEqual([])
+    })
+})
+
+describe('sorterSoknadGrupperEtterSignaturDato', () => {
+    it('sorterer nyeste sykmeldingUtskrevet øverst', () => {
+        const entries = [lagSokGruppe('gammel', '2026-01-01'), lagSokGruppe('ny', '2026-03-01')]
+
+        const resultat = sorterSoknadGrupperEtterSignaturDato(entries)
+
+        expect(resultat[0][0]).toBe('ny')
+        expect(resultat[1][0]).toBe('gammel')
+    })
+
+    it('plasserer gruppe uten sykmeldingUtskrevet sist', () => {
+        const entries = [lagSokGruppe('uten-dato', null), lagSokGruppe('med-dato', '2026-02-01')]
+
+        const resultat = sorterSoknadGrupperEtterSignaturDato(entries)
+
+        expect(resultat[0][0]).toBe('med-dato')
+        expect(resultat[1][0]).toBe('uten-dato')
+    })
+
+    it('returnerer tom liste uendret', () => {
+        expect(sorterSoknadGrupperEtterSignaturDato([])).toEqual([])
+    })
+})
