+++ vitest.config.ts
@@ -1,0 +1,6 @@
+import { defineConfig } from 'vitest/config'
+
+export default defineConfig({
+    envDir: '/tmp/vitest-env-dir',
+    test: {},
+})
